Tests are updated
