+++ apps/chat-e2e/src/tests/compareMode.test.ts
@@ -1282,7 +1282,7 @@ dialTest(
     dataInjector,
     compare,
     conversations,
-    chatBar,
+    header,
     chatHeader,
     compareConversation,
     conversationDropdownMenu,
@@ -1347,7 +1347,7 @@ dialTest(
     await dialTest.step(
       'Create new chat and verify Compare mode is closed',
       async () => {
-        await chatBar.createNewConversation();
+        await header.createNewConversation();
         await compare.waitForState({ state: 'hidden' });
       },
     );
