+++ apps/chat-e2e/src/tests/defaultModelSettings.test.ts
@@ -38,7 +38,7 @@ dialTest(
     'Addon icon is set in recent and selected list on default screen for new chat',
   async ({
     dialHomePage,
-    chatBar,
+    header,
     conversations,
     recentEntities,
     entitySettings,
@@ -66,7 +66,7 @@ dialTest(
         await dialHomePage.waitForPageLoaded({
           isNewConversationVisible: true,
         });
-        await chatBar.createNewConversation();
+        await header.createNewConversation();
 
         const todayConversations = await conversations.getTodayConversations();
         expect
@@ -213,7 +213,7 @@ dialTest(
     `It's impossible to send a message with spaces only`,
   async ({
     dialHomePage,
-    chatBar,
+    header,
     talkToSelector,
     chat,
     sendMessage,
@@ -319,7 +319,7 @@ dialTest(
     await dialTest.step(
       'Create new conversation and verify previous model is preselected and highlighted',
       async () => {
-        await chatBar.createNewConversation();
+        await header.createNewConversation();
         const modelBorderColors = await talkToEntities
           .getTalkToEntity(nonDefaultModel)
           .getAllBorderColors();
@@ -407,7 +407,7 @@ dialTest(
   'Recent "Talk to" list is updated',
   async ({
     dialHomePage,
-    chatBar,
+    header,
     chat,
     talkToSelector,
     marketplacePage,
@@ -424,7 +424,7 @@ dialTest(
       MockedChatApiResponseBodies.simpleTextBody,
     );
     await chat.sendRequestWithButton('test message');
-    await chatBar.createNewConversation();
+    await header.createNewConversation();
     const modelBorderColors = await talkToEntities
       .getTalkToEntity(nonDefaultModel)
       .getAllBorderColors();
